requirements.txt
Django>=4.2.3,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2-binary==2.9.6
drf-spectacular==0.26.3
Pillow==10.0.0